@@ -20,15 +20,15 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^8.1.0",
-    "@angular/common": "^8.1.0",
-    "@angular/compiler": "^8.1.0",
-    "@angular/core": "^8.1.0",
-    "@angular/forms": "^8.1.0",
-    "@angular/platform-browser": "^8.1.0",
-    "@angular/platform-browser-dynamic": "^8.1.0",
-    "@angular/router": "^8.1.0",
-    "@ngx-translate/core": "^11.0.1",
+    "@angular/animations": "^9.1.0",
+    "@angular/common": "^9.1.0",
+    "@angular/compiler": "^9.1.0",
+    "@angular/core": "^9.1.0",
+    "@angular/forms": "^9.1.0",
+    "@angular/platform-browser": "^9.1.0",
+    "@angular/platform-browser-dynamic": "^9.1.0",
+    "@angular/router": "^9.1.0",
+    "@ngx-translate/core": "^12.0.0",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/keyboard": "^5.0.0",
     "@ionic-native/splash-screen": "^5.0.0",
@@ -42,40 +42,39 @@
     "cordova-plugin-splashscreen": "^5.0.2",
     "cordova-plugin-statusbar": "^2.4.1",
     "cordova-plugin-whitelist": "^1.3.3",
-    "@ionic/angular": "^4.0.0",
-    "rxjs": "^6.5.2",
-    "zone.js": "^0.9.1"
+    "@ionic/angular": "^5.0.0",
+    "rxjs": "^6.5.4",
+    "tslib": "^1.10.1",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
     "cordova": "^9.0.0",
-    "@angular-builders/jest": "^8.0.4",
-    "@angular/cli": "~8.1.0",
-    "@angular/compiler-cli": "^8.1.0",
-    "@angular/language-service": "^8.1.0",
-    "@angular-devkit/build-angular": "^0.801.0",
+    "@angular-builders/jest": "^8.3.2",
+    "@angular/cli": "~9.1.0",
+    "@angular/compiler-cli": "^9.1.0",
+    "@angular/language-service": "^9.1.0",
+    "@angular-devkit/build-angular": "^0.901.0",
     "@angularclass/hmr": "^2.1.3",
-    "@biesbjerg/ngx-translate-extract": "^2.3.4",
+    "@biesbjerg/ngx-translate-extract": "^4.2.0",
     "@ngx-rocket/scripts": "^4.0.0",
     "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "^2.0.3",
-      "@types/jest": "^24.0.12",
-    "@types/node": "^10.9.0",
-    "codelyzer": "^5.1.0",
+      "@types/jest": "^25.1.2",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "htmlhint": "^0.11.0",
-    "https-proxy-agent": "^2.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^24.7.1",
-    "jest-preset-angular": "^7.1.0",
-    "protractor": "~5.4.0",
-    "puppeteer": "^1.17.0",
-    "stylelint": "~10.1.0",
-    "stylelint-config-recommended-scss": "~3.3.0",
-    "stylelint-config-standard": "~18.3.0",
-    "stylelint-scss": "~3.9.0",
-    "ts-node": "^8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.4.0"
+    "https-proxy-agent": "^5.0.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.1",
+    "jest": "^25.1.0",
+    "protractor": "~5.4.3",
+    "stylelint": "~13.2.1",
+    "stylelint-config-recommended-scss": "~4.2.0",
+    "stylelint-config-standard": "~20.0.0",
+    "stylelint-scss": "~3.16.0",
+    "ts-node": "^8.8.1",
+    "tslint": "~5.20.1",
+    "typescript": "~3.8.3"
   },
   "cordova": {
     "platforms": [
