@@ -25,8 +25,8 @@ describe('ErrorHandlerInterceptor', () => {
       }]
     });
 
-    http = TestBed.get(HttpClient);
-    httpMock = TestBed.get(HttpTestingController as Type<HttpTestingController>);
+    http = TestBed.inject(HttpClient);
+    httpMock = TestBed.inject(HttpTestingController as Type<HttpTestingController>);
   });
 
   afterEach(() => {
