@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { extract } from '@app/core';
+import { extract } from '@app/i18n';
 import { Shell } from '@app/shell/shell.service';
 import { AboutComponent } from './about.component';
 
