@@ -2,11 +2,14 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { TranslateModule } from '@ngx-translate/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ReactiveFormsModule } from '@angular/forms';
-import { FormsModule } from '@angular/forms';
 import { IonicModule, LoadingController, Platform } from '@ionic/angular';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 
-import { CoreModule } from '@app/core';
+import { CoreModule } from '@core';
+import { AuthenticationService, CredentialsService } from '@app/auth';
+import { MockAuthenticationService } from '@app/auth/authentication.service.mock';
+import { MockCredentialsService } from '@app/auth/credentials.service.mock';
+import { I18nModule } from '@app/i18n';
 import { LoginComponent } from './login.component';
 
 describe('LoginComponent', () => {
@@ -17,16 +20,18 @@ describe('LoginComponent', () => {
     TestBed.configureTestingModule({
       imports: [
         IonicModule.forRoot(),
-        FormsModule,
         RouterTestingModule,
         TranslateModule.forRoot(),
+        I18nModule,
         ReactiveFormsModule,
         CoreModule
       ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
       providers: [
+        { provide: AuthenticationService, useClass: MockAuthenticationService },
+        { provide: CredentialsService, useClass: MockCredentialsService },
         Platform,
-        LoadingController,
+        LoadingController
       ],
       declarations: [LoginComponent]
     })
