@@ -7,7 +7,8 @@ import { forkJoin, from } from 'rxjs';
 import { finalize } from 'rxjs/operators';
 
 import { environment } from '@env/environment';
-import { Logger, I18nService, AuthenticationService, untilDestroyed } from '@app/core';
+import { Logger, untilDestroyed } from '@core';
+import { AuthenticationService } from './authentication.service';
 
 const log = new Logger('Login');
 
@@ -28,7 +29,6 @@ export class LoginComponent implements OnInit, OnDestroy {
               private formBuilder: FormBuilder,
               private platform: Platform,
               private loadingController: LoadingController,
-              private i18nService: I18nService,
               private authenticationService: AuthenticationService) {
     this.createForm();
   }
@@ -59,17 +59,6 @@ export class LoginComponent implements OnInit, OnDestroy {
     });
   }
 
-  setLanguage(language: string) {
-    this.i18nService.language = language;
-  }
-
-  get currentLanguage(): string {
-    return this.i18nService.language;
-  }
-
-  get languages(): string[] {
-    return this.i18nService.supportedLanguages;
-  }
 
   get isWeb(): boolean {
     return !this.platform.is('cordova');
