@@ -3,8 +3,8 @@ import { CommonModule } from '@angular/common';
 import { TranslateModule } from '@ngx-translate/core';
 import { IonicModule } from '@ionic/angular';
 
-import { CoreModule } from '@app/core';
-import { SharedModule } from '@app/shared';
+import { CoreModule } from '@core';
+import { SharedModule } from '@shared';
 import { HomeRoutingModule } from './home-routing.module';
 import { HomeComponent } from './home.component';
 import { QuoteService } from './quote.service';
@@ -17,9 +17,6 @@ import { QuoteService } from './quote.service';
     SharedModule,
     IonicModule,
     HomeRoutingModule
-  ],
-  entryComponents: [
-    HomeComponent
   ],
   declarations: [
     HomeComponent
