@@ -2,9 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 import { ActionSheetController, AlertController, Platform } from '@ionic/angular';
-import { TextFieldTypes } from '@ionic/core';
 
-import { AuthenticationService, CredentialsService, I18nService } from '@app/core';
+import { AuthenticationService, CredentialsService } from '@app/auth';
 
 @Component({
   selector: 'app-settings',
@@ -16,12 +15,11 @@ export class SettingsComponent implements OnInit {
   constructor(private router: Router,
               private activatedRoute: ActivatedRoute,
               private translateService: TranslateService,
-              private platform: Platform,
-              private alertController: AlertController,
-              private actionSheetController: ActionSheetController,
               private authenticationService: AuthenticationService,
               private credentialsService: CredentialsService,
-              private i18nService: I18nService) { }
+              private platform: Platform,
+              private alertController: AlertController,
+              private actionSheetController: ActionSheetController) { }
 
   ngOnInit() { }
 
@@ -39,29 +37,4 @@ export class SettingsComponent implements OnInit {
       .subscribe(() => this.router.navigate(['/login'], { replaceUrl: true }));
   }
 
-  async changeLanguage() {
-    const alertController = await this.alertController.create({
-      header: this.translateService.instant('Change language'),
-      inputs: this.i18nService.supportedLanguages.map(language => ({
-        type: 'radio' as TextFieldTypes,
-        name: language,
-        label: language,
-        value: language,
-        checked: language === this.i18nService.language
-      })),
-      buttons: [
-        {
-          text: this.translateService.instant('Cancel'),
-          role: 'cancel'
-        },
-        {
-          text: this.translateService.instant('Ok'),
-          handler: language => {
-            this.i18nService.language = language;
-          }
-        }
-      ]
-    });
-    alertController.present();
-  }
 }
