@@ -4,6 +4,8 @@ import { TranslateModule } from '@ngx-translate/core';
 import { RouterModule } from '@angular/router';
 import { IonicModule } from '@ionic/angular';
 
+import { I18nModule } from '@app/i18n';
+import { AuthModule } from '@app/auth';
 import { ShellComponent } from './shell.component';
 
 @NgModule({
@@ -11,10 +13,9 @@ import { ShellComponent } from './shell.component';
     CommonModule,
     TranslateModule,
     IonicModule,
+    AuthModule,
+    I18nModule,
     RouterModule
-  ],
-  entryComponents: [
-    ShellComponent
   ],
   declarations: [
     ShellComponent
